# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MyCryptoCoin Payment Gateway API
  description: |
    MyCryptoCoin is the Stripe of crypto payments. Accept Bitcoin, Ethereum, USDT,
    and 20+ cryptocurrencies with a single integration. 0.5% per transaction, no
    hidden fees.

    ## Authentication
    All API requests require authentication via either:
    - **API Key**: Pass your key in the `X-API-Key` header
    - **Bearer Token**: Pass a JWT in the `Authorization: Bearer <token>` header

    API keys are prefixed:
    - `mcc_live_` for production
    - `mcc_test_` for sandbox/test mode

    ## Rate Limits
    Default rate limits are returned in response headers:
    - `X-RateLimit-Limit` — Requests allowed per window
    - `X-RateLimit-Remaining` — Requests remaining
    - `X-RateLimit-Reset` — Unix timestamp when the window resets

    ## Pagination
    List endpoints support cursor-based pagination:
    - `page` — Page number (default: 1)
    - `limit` — Items per page (default: 20, max: 100)

    Response includes `pagination` object with `total`, `page`, `limit`, `pages`.

    ## Errors
    All errors follow a consistent format:
    ```json
    {
      "error": {
        "code": "invalid_request",
        "message": "Human-readable description",
        "details": {}
      }
    }
    ```
  version: "1.0.0"
  contact:
    name: MyCryptoCoin Developer Support
    email: developers@mycrypto.co.in
    url: https://mycrypto.co.in/support
  license:
    name: Proprietary
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key prefixed with mcc_live_ or mcc_test_"
  headers:
    X-RateLimit-Limit:
      schema:
        type: integer
      description: Maximum requests per rate limit window
    X-RateLimit-Remaining:
      schema:
        type: integer
      description: Remaining requests in the current window
    X-RateLimit-Reset:
      schema:
        type: integer
      description: Unix timestamp when the rate limit window resets
    X-Request-Id:
      schema:
        type: string
        format: uuid
      description: Unique identifier for the request, useful for support
  schemas:
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          description: Total number of records
          example: 142
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Records per page
          example: 20
        pages:
          type: integer
          description: Total number of pages
          example: 8

    # ─── Error ────────────────────────────────────────────────────
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: invalid_request
            message:
              type: string
              description: Human-readable error message
              example: "The 'amount' field is required."
            details:
              type: object
              additionalProperties: true
              description: Additional context about the error
              example:
                field: amount
                reason: required

    # ─── Auth Schemas ─────────────────────────────────────────────
...
    Transaction:
      type: object
      properties:
        id:
          type: string
          example: "txn_q1w2e3r4t5"
        type:
          type: string
          enum:
            - payment_received
            - withdrawal
            - fee
            - auto_withdrawal
            - refund
          example: payment_received
        status:
          type: string
          enum: [pending, confirming, completed, failed]
          example: completed
        crypto:
          type: string
          example: USDT
        amount:
          type: string
          example: "99.990000"
        fee:
          type: string
          example: "0.499950"
        net_amount:
          type: string
          example: "99.490050"
        balance_after:
          type: string
          description: Wallet balance after this transaction
          example: "1245.380050"
        reference_id:
          type: string
          description: "Related payment or withdrawal ID"
          example: "pay_1a2b3c4d5e6f"
        reference_type:
          type: string
          enum: [payment, withdrawal, refund]
          example: payment
        description:
          type: string
          example: "Payment received - ORD-12345"
        tx_hash:
          type: string
          nullable: true
          example: "0xabc123def456..."
        created_at:
          type: string
          format: date-time
          example: "2026-03-19T10:45:30Z"
    TransactionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # ─── Webhook Schemas ──────────────────────────────────────────
...
    Webhook:
      type: object
      properties:
        id:
          type: string
          example: "whk_m1n2o3p4q5"
        url:
          type: string
          example: "https://yoursite.com/webhooks/mycryptocoin"
        events:
          type: array
          items:
            type: string
          example: ["payment.confirmed", "payment.failed", "withdrawal.completed"]
        description:
          type: string
          example: "Production payment notifications"
        active:
          type: boolean
          example: true
        secret:
          type: string
          description: "HMAC-SHA256 signing secret (shown only on creation)"
          example: "whsec_a1b2c3d4e5f6g7h8i9j0..."
        last_delivery:
          type: object
          nullable: true
          properties:
            event:
              type: string
              example: "payment.confirmed"
            status_code:
              type: integer
              example: 200
            delivered_at:
              type: string
              format: date-time
              example: "2026-03-19T10:45:30Z"
        created_at:
          type: string
          format: date-time
          example: "2026-03-19T08:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2026-03-19T10:45:30Z"
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          example: "https://yoursite.com/webhooks/v2/mycryptocoin"
        events:
          type: array
          items:
            type: string
          example: ["*"]
        description:
          type: string
          example: "Updated webhook"
        active:
          type: boolean
          example: true
...
  responses:
    BadRequest:
      description: Invalid request parameters
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: invalid_request
              message: "The 'amount' field must be greater than 0."
              details:
                field: amount
                reason: minimum
    Unauthorized:
      description: Authentication required or token expired
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: authentication_required
              message: "Invalid or expired authentication token."
...
    NotFound:
      description: Resource not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: not_found
              message: "Payment pay_1a2b3c4d5e6f not found."
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          $ref: '#/components/headers/X-RateLimit-Limit'
        X-RateLimit-Remaining:
          $ref: '#/components/headers/X-RateLimit-Remaining'
        X-RateLimit-Reset:
          $ref: '#/components/headers/X-RateLimit-Reset'
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: rate_limit_exceeded
              message: "Too many requests. Retry after 30 seconds."
              details:
                retry_after: 30
    InternalError:
      description: Internal server error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/X-Request-Id'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: internal_error
              message: "An unexpected error occurred. Please try again later."
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [Transactions]
      summary: List all transactions
      description: |
        Retrieve a unified, paginated list of all transactions including
        payments received, withdrawals, fees, and refunds.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [payment_received, withdrawal, fee, auto_withdrawal, refund]
          description: Filter by transaction type
        - name: crypto
          in: query
          schema:
            type: string
          description: Filter by cryptocurrency
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirming, completed, failed]
        - name: date_from
          in: query
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: List of transactions
          headers:
            X-RateLimit-Limit:
              $ref: '#/components/headers/X-RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/X-RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/X-RateLimit-Reset'
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

  # ═══════════════════════════════════════════════════════════════
  #  WEBHOOKS
  # ═══════════════════════════════════════════════════════════════
...
  /webhooks/{id}:
    put:
      operationId: updateWebhook
      tags: [Webhooks]
      summary: Update a webhook endpoint
      description: Update the URL, events, or active status of a webhook.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: deleteWebhook
      tags: [Webhooks]
      summary: Delete a webhook endpoint
      description: Permanently delete a webhook endpoint. This cannot be undone.
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook deleted
          headers:
            X-Request-Id:
              $ref: '#/components/headers/X-Request-Id'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Webhook whk_m1n2o3p4q5 deleted."
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...